requirements.txt
# LangChain Ecosystem Workshop Dependencies
# Install with: pip install -r requirements.txt

# Core LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# LangGraph - Stateful Workflows
langgraph>=0.2.0

# LangSmith - Observability
langsmith>=0.1.0

# LLM Providers
openai>=1.0.0
anthropic>=0.20.0

# Vector Stores
chromadb>=0.5.0

# Embeddings
sentence-transformers>=3.0.0

# Data Validation
pydantic>=2.0.0

# Async Support
aiohttp>=3.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
python-dotenv>=1.0.0
ipython>=8.0.0

# Optional: Web Interface
streamlit>=1.28.0

# Optional: API
fastapi>=0.115.0
uvicorn>=0.30.0
